refactor(cli): extract session-analysis/settings/packages handlers out of daemon.ts
Phase 1b. Settings and packages sections were fully decoupled from
daemon state (socket + isShuttingDown() only); session-analysis needed
runningSessions/sessionCloseMetadata (read-only) plus two closures
that stay owned by daemon.ts. daemon.ts: 2055 -> 1523 lines (37.6%
cut from the pre-refactor 2442 total). Confirmed the 13 full-suite
failures (cli-colors/setup/sandbox-config/copyBinaryAssets) are
pre-existing test-run flakiness, unrelated: identical failures
reproduce on unmodified main.
