+++ packages/cli/src/runner/daemon.ts
@@ -2,7 +2,7 @@ import { exec, type ChildProcess } from "node:child_process";
 import { promisify } from "node:util";
 
 const execAsync = promisify(exec);
-import { hostname, homedir } from "node:os";
+import { hostname } from "node:os";
 import { dirname, join } from "node:path";
 import { existsSync, rmSync } from "node:fs";
 import { forceKillTree, isShutdownMessage, requestChildShutdown, STOP_FILE_NAME } from "./process-kill.js";
@@ -36,7 +36,6 @@ import { ModelRegistry, ModelRuntime } from "@earendil-works/pi-coding-agent";
 import type { ServiceTriggerDef, ServiceSigilDef, TriggerSubscriptionEntry } from "@pizzapi/protocol";
 import { setLogComponent, logInfo, logWarn, logError } from "./logger.js";
 import { extractHookSummary } from "./hook-summary.js";
-import { sanitizeConfigForUI, restoreMaskedServerEntry, findRenamedServerMatch, MASK_SENTINEL, validateProviderOverridesSection, mergeProviderOverridesSection } from "./daemon-config-sanitize.js";
 import { defaultStatePath, acquireStateAndIdentity, releaseStateLock } from "./runner-state.js";
 import { normalizeLoopbackHost } from "../relay-url.js";
 import { startUsageRefreshLoop, stopUsageRefreshLoop } from "./runner-usage-cache.js";
@@ -56,6 +55,9 @@ import { registerPluginsHandlers } from "./daemon-handlers/plugins.js";
 import { registerSandboxHandlers } from "./daemon-handlers/sandbox.js";
 import { registerModelsHandlers } from "./daemon-handlers/models.js";
 import { registerUsageHandlers } from "./daemon-handlers/usage.js";
+import { registerSessionAnalysisHandlers } from "./daemon-handlers/session-analysis.js";
+import { registerSettingsHandlers } from "./daemon-handlers/settings.js";
+import { registerPackagesHandlers } from "./daemon-handlers/packages.js";
 
 // Re-export migration from shared module — used on daemon startup
 import { migrateAgentDir } from "../migrations.js";
@@ -1500,551 +1502,17 @@ export async function runDaemon(_args: string[] = []): Promise<number> {
         registerModelsHandlers(socket, () => isShuttingDown, listConfiguredModels);
         registerUsageHandlers(socket, () => isShuttingDown);
 
-        // ── Session Analysis ──────────────────────────────────────────
-
-        socket.on("analyze_session", async (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data.requestId ?? "";
-            try {
-                const sessionId = data.sessionId;
-                if (!sessionId || typeof sessionId !== "string") {
-                    socket.emit("analyze_session_error", {
-                        requestId,
-                        error: "Missing sessionId parameter",
-                    });
-                    return;
-                }
-                const sessionMetadata = sessionCloseMetadata.get(sessionId);
-                const sessionsRootDir = join(resolveConfiguredAgentDir(sessionMetadata?.cwd), "sessions");
-                const sessionFile = runningSessions.get(sessionId)?.sessionFile
-                    ?? sessionMetadata?.sessionFile
-                    ?? await findSessionPathById(sessionsRootDir, sessionId);
-                if (!sessionFile) {
-                    socket.emit("analyze_session_error", {
-                        requestId,
-                        error: "Session file not found for " + sessionId,
-                    });
-                    return;
-                }
-                const MAX_ANALYSIS_FILE_SIZE = 50 * 1024 * 1024; // 50 MB
-                const transcriptFile = Bun.file(sessionFile);
-                if (!await transcriptFile.exists()) {
-                    socket.emit("analyze_session_error", {
-                        requestId,
-                        error: "Session file does not exist: " + sessionFile,
-                    });
-                    return;
-                }
-                if (transcriptFile.size > MAX_ANALYSIS_FILE_SIZE) {
-                    socket.emit("analyze_session_error", {
-                        requestId,
-                        error: `Session file too large for analysis (${Math.round(transcriptFile.size / 1024 / 1024)} MB, max ${MAX_ANALYSIS_FILE_SIZE / 1024 / 1024} MB)`,
-                    });
-                    return;
-                }
-                const { parseJsonlEntries } = await import("../session-analysis/parser.js");
-                const { reconstructContext } = await import("../session-analysis/analyzer.js");
-                const content = await transcriptFile.text();
-                const { entries } = parseJsonlEntries(content);
-                const leafId = entries.findLast((e: any) => e.id)?.id ?? "root";
-                const analysis = reconstructContext(
-                    entries,
-                    leafId,
-                    await getContextWindowsForAnalysis(sessionMetadata?.cwd),
-                );
-                socket.emit("analyze_session_data", { requestId, data: analysis });
-            } catch (e: any) {
-                socket.emit("analyze_session_error", {
-                    requestId,
-                    error: e.message ?? "Unknown error",
-                });
-            }
-        });
-
-        // ── Settings ───────────────────────────────────────────────────────
-
-        // sanitizeConfigForUI is imported from ./daemon-config-sanitize.js
-
-        socket.on("settings_get_config", async (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data?.requestId;
-            try {
-                const { loadGlobalConfig: loadGlobal } = await import("../config.js");
-                const globalConfig = loadGlobal();
-
-                // Also read settings.json (TUI preferences)
-                const settingsPath = join(homedir(), ".pizzapi", "settings.json");
-                let tuiSettings: Record<string, unknown> = {};
-                try {
-                    const { readFileSync, existsSync } = await import("fs");
-                    if (existsSync(settingsPath)) {
-                        tuiSettings = JSON.parse(readFileSync(settingsPath, "utf-8"));
-                    }
-                } catch {
-                    // settings.json may not exist yet
-                }
-
-                // Also read ~/.pizzapi/AGENTS.md
-                const agentsMdPath = join(homedir(), ".pizzapi", "AGENTS.md");
-                let agentsMd = "";
-                try {
-                    const { readFileSync, existsSync } = await import("fs");
-                    if (existsSync(agentsMdPath)) {
-                        agentsMd = readFileSync(agentsMdPath, "utf-8");
-                    }
-                } catch {
-                    // AGENTS.md may not exist yet
-                }
-
-                // Strip sensitive fields before sending to UI
-                const sanitizedConfig = sanitizeConfigForUI(globalConfig as Record<string, unknown>);
-                socket.emit("file_result", {
-                    requestId,
-                    ok: true,
-                    config: sanitizedConfig,
-                    tuiSettings,
-                    agentsMd,
-                });
-            } catch (err) {
-                socket.emit("file_result", {
-                    requestId,
-                    ok: false,
-                    message: err instanceof Error ? err.message : String(err),
-                });
-            }
-        });
-
-        socket.on("settings_update_section", async (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data?.requestId;
-            const section = data?.section;
-            const value = data?.value;
-            try {
-                if (!section || typeof section !== "string") {
-                    socket.emit("file_result", { requestId, ok: false, message: "Missing section name" });
-                    return;
-                }
-
-                // Handle AGENTS.md separately — it's a standalone file, not JSON config
-                if (section === "agentsMd") {
-                    const agentsMdPath = join(homedir(), ".pizzapi", "AGENTS.md");
-                    const { writeFileSync, chmodSync: chmodSyncAgents, mkdirSync, existsSync } = await import("fs");
-                    const dir = join(homedir(), ".pizzapi");
-                    if (!existsSync(dir)) mkdirSync(dir, { recursive: true, mode: 0o700 });
-                    const content = typeof value === "string" ? value : "";
-                    writeFileSync(agentsMdPath, content, { encoding: "utf-8", mode: 0o600 });
-                    chmodSyncAgents(agentsMdPath, 0o600); // tighten permissions on pre-existing files
-                    socket.emit("file_result", {
-                        requestId,
-                        ok: true,
-                        saved: true,
-                        message: "AGENTS.md saved. Changes apply on next session start.",
-                    });
-                    return;
-                }
-
-                // Sections that go into settings.json (TUI preferences)
-                const tuiSections = new Set(["tuiPreferences", "models"]);
-
-                if (tuiSections.has(section)) {
-                    // Read/merge/write settings.json
-                    const settingsPath = join(homedir(), ".pizzapi", "settings.json");
-                    const { readFileSync, writeFileSync, chmodSync: chmodSyncSettings, existsSync, mkdirSync } = await import("fs");
-                    const dir = join(homedir(), ".pizzapi");
-                    if (!existsSync(dir)) mkdirSync(dir, { recursive: true, mode: 0o700 });
-                    let existing: Record<string, unknown> = {};
-                    try {
-                        if (existsSync(settingsPath)) {
-                            existing = JSON.parse(readFileSync(settingsPath, "utf-8"));
-                        }
-                    } catch { /* start fresh */ }
-
-                    // Merge TUI settings at top level
-                    if (value && typeof value === "object" && !Array.isArray(value)) {
-                        Object.assign(existing, value);
-                    }
-                    writeFileSync(settingsPath, JSON.stringify(existing, null, 2), { encoding: "utf-8", mode: 0o600 });
-                    chmodSyncSettings(settingsPath, 0o600); // tighten permissions on pre-existing files
-                    socket.emit("file_result", {
-                        requestId,
-                        ok: true,
-                        saved: true,
-                        message: "TUI settings saved. Changes apply on next session start.",
-                    });
-                } else {
-                    // All other sections go into config.json
-                    const { saveGlobalConfig: saveGlobal, loadGlobalConfig: loadGlobal } = await import("../config.js");
-
-                    // Map section names to config.json keys
-                    const sectionToConfigKey: Record<string, string> = {
-                        mcpServers: "mcpServers",
-                        hooks: "hooks",
-                        sandbox: "sandbox",
-                        webSearch: "providerSettings",
-                        toolSearch: "toolSearch",
-                        security: "_security",        // virtual — handled specially
-                        envVars: "_envVars",          // virtual — handled specially
-                        systemPrompt: "_systemPrompt", // virtual — handled specially
-                    };
-
-                    const configKey = sectionToConfigKey[section] ?? section;
-                    const existing = loadGlobal();
-
-                    if (section === "security") {
-                        const v = value as any;
-                        const updates: Record<string, any> = {};
-                        if (v?.allowProjectHooks !== undefined) updates.allowProjectHooks = v.allowProjectHooks;
-                        if (v?.trustedPlugins !== undefined) updates.trustedPlugins = v.trustedPlugins;
-                        saveGlobal(updates);
-                    } else if (section === "systemPrompt") {
-                        const v = value as any;
-                        const updates: Record<string, any> = {};
-                        if (v?.appendSystemPrompt !== undefined) updates.appendSystemPrompt = v.appendSystemPrompt;
-                        if (v?.builtinSystemPrompt !== undefined) updates.builtinSystemPrompt = v.builtinSystemPrompt;
-                        if (v?.sendAgentsMd !== undefined) updates.sendAgentsMd = v.sendAgentsMd;
-                        if (v?.skills !== undefined) updates.skills = v.skills;
-                        saveGlobal(updates);
-                    } else if (section === "envVars") {
-                        // Env vars are stored in a custom key in config.json.
-                        // Restore any masked ("***") sentinel values from the on-disk config
-                        // so we don't overwrite real secrets with the placeholder.
-                        const MASK_SENTINEL = "***";
-                        const existingOverrides = ((existing as any).envOverrides ?? {}) as Record<string, string>;
-                        const incomingOverrides = (value ?? {}) as Record<string, string>;
-                        const restoredOverrides: Record<string, string> = { ...incomingOverrides };
-                        for (const [k, v] of Object.entries(incomingOverrides)) {
-                            if (v === MASK_SENTINEL && typeof existingOverrides[k] === "string") {
-                                restoredOverrides[k] = existingOverrides[k];
-                            }
-                        }
-                        const updates: Record<string, any> = { envOverrides: restoredOverrides };
-                        saveGlobal(updates);
-                    } else if (section === "providerOverrides") {
-                        // Per-provider overrides (system prompt, AGENTS.md, MCP disable
-                        // list) go into providerSettings.<provider>.overrides.
-                        const validationErrors = validateProviderOverridesSection(value);
-                        if (validationErrors.length > 0) {
-                            socket.emit("file_result", {
-                                requestId,
-                                ok: false,
-                                message: `Invalid provider overrides:\n${validationErrors.join("\n")}`,
-                            });
-                            return;
-                        }
-                        const ps = mergeProviderOverridesSection(
-                            (existing as any).providerSettings,
-                            (value ?? {}) as Record<string, unknown>,
-                        );
-                        saveGlobal({ providerSettings: ps } as any);
-                    } else if (section === "webSearch") {
-                        // Web search config goes into providerSettings
-                        const v = value as any;
-                        const ps = (existing as any).providerSettings ?? {};
-                        if (v?.anthropic?.webSearch) {
-                            ps.anthropic = { ...ps.anthropic, webSearch: v.anthropic.webSearch };
-                        }
-                        if (v?.["ollama-cloud"]?.webSearch) {
-                            ps["ollama-cloud"] = { ...ps["ollama-cloud"], webSearch: v["ollama-cloud"].webSearch };
-                        }
-                        saveGlobal({ providerSettings: ps } as any);
-                    } else if (section === "toolSearch") {
-                        if (value != null && (typeof value !== "object" || Array.isArray(value))) {
-                            socket.emit("file_result", {
-                                requestId,
-                                ok: false,
-                                message: "toolSearch must be a JSON object",
-                            });
-                            return;
-                        }
-
-                        const toolSearch = (value ?? {}) as Record<string, unknown>;
-                        const errors: string[] = [];
-                        const enabled = toolSearch.enabled;
-                        const tokenThreshold = toolSearch.tokenThreshold;
-                        const maxResults = toolSearch.maxResults;
-                        const keepLoadedTools = toolSearch.keepLoadedTools;
-
-                        if (enabled !== undefined && typeof enabled !== "boolean") {
-                            errors.push('"enabled" must be a boolean');
-                        }
-                        if (
-                            tokenThreshold !== undefined &&
-                            (typeof tokenThreshold !== "number" || !Number.isFinite(tokenThreshold) || tokenThreshold < 0)
-                        ) {
-                            errors.push('"tokenThreshold" must be a finite number >= 0');
-                        }
-                        if (
-                            maxResults !== undefined &&
-                            (typeof maxResults !== "number" || !Number.isFinite(maxResults) || maxResults < 1)
-                        ) {
-                            errors.push('"maxResults" must be a finite number >= 1');
-                        }
-                        if (keepLoadedTools !== undefined && typeof keepLoadedTools !== "boolean") {
-                            errors.push('"keepLoadedTools" must be a boolean');
-                        }
-                        if (errors.length > 0) {
-                            socket.emit("file_result", {
-                                requestId,
-                                ok: false,
-                                message: `Invalid Tool Search config:\n${errors.join("\n")}`,
-                            });
-                            return;
-                        }
-
-                        saveGlobal({ toolSearch: value as any });
-                    } else if (section === "mcpServers") {
-                        // Validate MCP server config before saving
-                        if (value != null && (typeof value !== "object" || Array.isArray(value))) {
-                            socket.emit("file_result", {
-                                requestId,
-                                ok: false,
-                                message: "mcpServers must be a JSON object (Record<string, ServerEntry>)",
-                            });
-                            return;
-                        }
-                        const servers = (value ?? {}) as Record<string, any>;
-                        const errors: string[] = [];
-                        for (const [name, entry] of Object.entries(servers)) {
-                            if (entry == null || typeof entry !== "object" || Array.isArray(entry)) {
-                                errors.push(`"${name}": must be an object`);
-                                continue;
-                            }
-                            const hasCommand = typeof entry.command === "string" && entry.command.trim() !== "";
-                            const hasUrl = typeof entry.url === "string" && entry.url.trim() !== "";
-                            if (!hasCommand && !hasUrl) {
-                                errors.push(`"${name}": must have a "command" (stdio) or "url" (http) field`);
-                            }
-                        }
-                        if (errors.length > 0) {
-                            socket.emit("file_result", {
-                                requestId,
-                                ok: false,
-                                message: `Invalid MCP server config:\n${errors.join("\n")}`,
-                            });
-                            return;
-                        }
-
-                        // The settings API masks sensitive env var and header values (e.g. tokens/keys)
-                        // with MASK_SENTINEL ("***") before sending them to the UI so they aren't
-                        // exposed in transit.  When the UI sends the config back on save, those
-                        // masked values must NOT be written to disk — restoreMaskedServerEntry()
-                        // substitutes the original on-disk secret for each key still carrying the
-                        // sentinel.
-                        //
-                        // Renames are handled heuristically by findRenamedServerMatch() below:
-                        // when an incoming entry has masked secrets but no on-disk entry under its
-                        // name, we look for a deleted entry whose env/header keys would supply the
-                        // sentinels.  This survives the user editing command/url/args in the same
-                        // save.  Truly ambiguous cases (multiple plausible renames) still fall back
-                        // to writing the sentinel — visible and recoverable.
-                        const existingMcpServers = ((existing as any).mcpServers ?? {}) as Record<string, any>;
-
-                        // Identify deleted servers to heuristically match renames
-                        const incomingNames = new Set(Object.keys(servers));
-                        const deletedServers = Object.entries(existingMcpServers)
-                            .filter(([name]) => !incomingNames.has(name))
-                            .map(([_name, srv]) => srv)
-                            .filter((srv) => srv && typeof srv === "object");
-
-                        const mergedServers: Record<string, any> = {};
-                        for (const [name, entry] of Object.entries(servers)) {
-                            if (entry && typeof entry === "object") {
-                                let existingEntry = existingMcpServers[name];
-                                if (!existingEntry) {
-                                    existingEntry = findRenamedServerMatch(entry as Record<string, unknown>, deletedServers);
-                                }
-
-                                mergedServers[name] = restoreMaskedServerEntry(
-                                    entry as Record<string, unknown>,
-                                    existingEntry,
-                                );
-                            } else {
-                                mergedServers[name] = entry;
-                            }
-                        }
-                        saveGlobal({ mcpServers: mergedServers } as any);
-                    } else if (section === "mcp") {
-                        // mcp.servers[] (preferred array format) — restore masked sentinel values
-                        // before writing to disk.  We look up each server by its `name` field in
-                        // the on-disk array so we can restore the original secret.
-                        //
-                        // Renames in the array format are likewise resolved via
-                        // findRenamedServerMatch() against deleted entries.
-                        const incomingMcp = (value ?? {}) as { servers?: any[] };
-                        const existingMcp = ((existing as any).mcp ?? {}) as { servers?: any[] };
-
-                        // Build name → entry map for O(1) lookup against the on-disk array.
-                        const existingByName = new Map<string, Record<string, unknown>>();
-                        if (Array.isArray(existingMcp.servers)) {
-                            for (const s of existingMcp.servers) {
-                                if (s && typeof s === "object" && typeof (s as any).name === "string") {
-                                    existingByName.set((s as any).name as string, s as Record<string, unknown>);
-                                }
-                            }
-                        }
-
-                        // Identify deleted servers to heuristically match renames
-                        const incomingNamesArray = new Set(
-                            Array.isArray(incomingMcp.servers)
-                                ? incomingMcp.servers
-                                      .map((s: any) => s && typeof s === "object" ? s.name : undefined)
-                                      .filter((n): n is string => typeof n === "string")
-                                : []
-                        );
-
-                        const deletedServersArray: Record<string, unknown>[] = [];
-                        if (Array.isArray(existingMcp.servers)) {
-                            for (const srv of existingMcp.servers) {
-                                if (srv && typeof srv === "object" && typeof (srv as any).name === "string") {
-                                    if (!incomingNamesArray.has((srv as any).name)) {
-                                        deletedServersArray.push(srv as Record<string, unknown>);
-                                    }
-                                }
-                            }
-                        }
-
-                        const mergedMcpServers: any[] = Array.isArray(incomingMcp.servers)
-                            ? incomingMcp.servers.map((entry: any) => {
-                                  if (!entry || typeof entry !== "object") return entry;
-
-                                  let existingEntry: Record<string, unknown> | undefined = undefined;
-                                  if (typeof entry.name === "string") {
-                                      existingEntry = existingByName.get(entry.name);
-                                      if (!existingEntry) {
-                                          existingEntry = findRenamedServerMatch(entry as Record<string, unknown>, deletedServersArray);
-                                      }
-                                  }
-
-                                  return restoreMaskedServerEntry(
-                                      entry as Record<string, unknown>,
-                                      existingEntry,
-                                  );
-                              })
-                            : [];
-
-                        saveGlobal({ mcp: { ...incomingMcp, servers: mergedMcpServers } } as any);
-                    } else {
-                        // Direct key mapping
-                        saveGlobal({ [configKey]: value } as any);
-                    }
-
-                    // Reload and return the updated config — mask secrets before sending to browser
-                    const updatedConfig = sanitizeConfigForUI(loadGlobal() as Record<string, unknown>);
-                    const isMcpSection = section === "mcpServers" || section === "mcp" || section === "toolSearch";
-                    const reloadHint = isMcpSection
-                        ? "MCP server config saved. Active sessions can run /mcp reload to pick up changes."
-                        : "Settings saved. Changes apply on next session start.";
-                    socket.emit("file_result", {
-                        requestId,
-                        ok: true,
-                        saved: true,
-                        config: updatedConfig,
-                        message: reloadHint,
-                        reloadHint: isMcpSection,
-                    });
-                }
-            } catch (err) {
-                socket.emit("file_result", {
-                    requestId,
-                    ok: false,
-                    message: err instanceof Error ? err.message : String(err),
-                });
-            }
-        });
-
-        // ── Package management ────────────────────────────────────────────────
-
-        async function withPackageManager<T extends Record<string, unknown>>(
-            socket: any,
-            requestId: string,
-            fn: (pm: any) => Promise<T>,
-        ): Promise<void> {
-            try {
-                const { DefaultPackageManager, SettingsManager, getAgentDir } = await import("@earendil-works/pi-coding-agent");
-                const agentDir = getAgentDir();
-                const cwd = process.cwd();
-                const settingsManager = SettingsManager.create(cwd, agentDir);
-                const packageManager = new DefaultPackageManager({ cwd, agentDir, settingsManager });
-                const result = await fn(packageManager);
-                socket.emit("file_result", { requestId, ok: true, ...result });
-            } catch (err) {
-                socket.emit("file_result", { requestId, ok: false, message: err instanceof Error ? err.message : String(err) });
-            }
-        }
-
-        socket.on("packages_list", async (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data?.requestId;
-            await withPackageManager(socket, requestId, async (packageManager) => {
-                const packages = packageManager.listConfiguredPackages();
-                return { packages, message: `Found ${packages.length} package(s)` };
-            });
-        });
-
-        socket.on("packages_install", async (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data?.requestId;
-            const source = data?.source;
-            const isLocal = data?.local === true;
-
-            if (!source || typeof source !== "string") {
-                socket.emit("file_result", {
-                    requestId,
-                    ok: false,
-                    message: "Missing or invalid package source",
-                });
-                return;
-            }
-
-            await withPackageManager(socket, requestId, async (packageManager) => {
-                await packageManager.install(source, { local: isLocal });
-                const added = packageManager.addSourceToSettings(source, { local: isLocal });
-                if (!added) {
-                    throw new Error(`Failed to add package ${source} to settings`);
-                }
-                return { message: `Package ${source} installed successfully` };
-            });
-        });
-
-        socket.on("packages_remove", async (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data?.requestId;
-            const source = data?.source;
-            const isLocal = data?.local === true;
-
-            if (!source || typeof source !== "string") {
-                socket.emit("file_result", {
-                    requestId,
-                    ok: false,
-                    message: "Missing or invalid package source",
-                });
-                return;
-            }
-
-            await withPackageManager(socket, requestId, async (packageManager) => {
-                await packageManager.remove(source, { local: isLocal });
-                const removed = packageManager.removeSourceFromSettings(source, { local: isLocal });
-                if (!removed) {
-                    throw new Error(`Package ${source} not found in settings`);
-                }
-                return { message: `Package ${source} removed successfully` };
-            });
-        });
-
-        socket.on("packages_update", async (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data?.requestId;
-            const source = data?.source; // Optional: update specific package
-
-            await withPackageManager(socket, requestId, async (packageManager) => {
-                await packageManager.update(source);
-                return {
-                    message: source
-                        ? `Package ${source} updated successfully`
-                        : "All packages updated successfully",
-                };
-            });
-        });
+        // ── Session analysis / settings / packages handlers ────────────────
+        registerSessionAnalysisHandlers(
+            socket,
+            () => isShuttingDown,
+            runningSessions,
+            sessionCloseMetadata,
+            resolveConfiguredAgentDir,
+            getContextWindowsForAnalysis,
+        );
+        registerSettingsHandlers(socket, () => isShuttingDown);
+        registerPackagesHandlers(socket, () => isShuttingDown);
 
         // ── Error handling ────────────────────────────────────────────────
